move addBook slice out of /slices for consistency
